# backend/task1/openapi.yml
openapi: 3.0.0
info:
  title: Task 1
  description: Tasks API
  version: 0.1.0
servers:
  - url: http://localhost:8000/v1
    description: Development server
paths:
  /tasks:
    get:
      summary: List all tasks
      operationId: list_tasks
      tags:
        - Tasks
      parameters:
        - name: cursor
          in: query
          description: Cursor
          schema:
            type: string
        - name: page_size
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTasksResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add a new task
      operationId: add_task
      tags:
        - Tasks
      requestBody:
        description: Task to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTaskRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddTaskResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}:
    get:
      summary: Retrieve a specific task
      operationId: get_task
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTaskResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a specific task
      operationId: update_task
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
      requestBody:
        description: Task to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTaskResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a specific task
      operationId: delete_task
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTaskResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
    ListTasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        next_cursor:
          type: string
    AddTaskRequest:
      type: object
      properties:
        title:
          type: string
    AddTaskResponse:
      type: object
      properties:
        task:
          $ref: '#/components/schemas/Task'
    GetTaskResponse:
      type: object
      properties:
        task:
          $ref: '#/components/schemas/Task'
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
    UpdateTaskResponse:
      type: object
      properties:
        task:
          $ref: '#/components/schemas/Task'
    DeleteTaskResponse:
      type: object
      properties:
        task:
          $ref: '#/components/schemas/Task'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string